+++ src/NetCore/Web/Api/wwwroot/scripts/api/webgis.l10n.de.js
@@ -294,6 +294,8 @@
     "sketch-ortho-off": "Orthogonal-Modus beenden",
     "sketch-trace-off": "Trace-Modus beenden",
     "sketch-fan-off": "Fan-Modus beenden",
+    "sections": "Abschnitte",
+    "section": "Abschnitt",
 
     // query combos
     "query-choose-category": "Kategorie wählen (optional)",
@@ -328,5 +330,6 @@
     "show-markers-on-new-queries": "Marker bei neuen Abfragen anzeigen",
     "show-markers-on-new-queries-info": "Wenn diese Methode angewendet wird, werden bei jeder neuen Abfrage automatisch Kartenmarken in der Karte angezeigt. Das Defaultverhalten ist, das Marker automatisch dargestellt werden.",
     "select-new-query-results": "Neue Abfragen in der Karte auswählen/selektieren",
-    "select-new-query-results-info": "Wenn diese Methode angewendet wird, werden Abfrageergebnisse in der Karte automatisch selektiert dargestellt."
+    "select-new-query-results-info": "Wenn diese Methode angewendet wird, werden Abfrageergebnisse in der Karte automatisch selektiert dargestellt.",
+    "query-layer-not-visible-notification": "Der abgefrage Layer ist in der Karte ausgeblendet. Hier Klicken zum Einblenden."
 };