+++ src/NetCore/Web/Api/wwwroot/scripts/api/webgis.options.js
@@ -93,6 +93,8 @@ webgis.usability = {
 
     useAdvancedKeyShortcutHandling: false,
 
+    showQueryLayerNotVisbleNotification: false,
+
     tsp: {
         allowOrderFeatures: false,
         maxFeatures: 50
@@ -104,7 +106,8 @@ webgis.usability = {
         displayMetadata: {
             default: true,
             geocodes: false
-        }
+        },
+        selectFirstOnEnter: false
     },
 
     sketchTools: {