+++ src/NetStandard/E.Standard.WebMapping.Core/Abstraction/IServiceAttachmentProvider.cs
@@ -1,0 +1,8 @@
+﻿using System.Threading.Tasks;
+
+namespace E.Standard.WebMapping.Core.Abstraction;
+
+public interface IServiceAttachmentProvider 
+{
+    Task<byte[]> GetServiceAttachementData(IRequestContext requestContext, string attachmentUri);
+}