+++ src/NetStandard/E.Standard.WebMapping.GeoServices/ArcServer/Rest/Extensions/AgsErrorResponseExtensions.cs
@@ -10,9 +10,25 @@ namespace E.Standard.WebMapping.GeoServices.ArcServer.Rest.Extensions;
 
 static class AgsErrorResponseExtensions
 {
+    //public static string UrlEncodePassword(this string password)
+    //{
+    //    if (password != null && password.IndexOfAny("+/=&".ToCharArray()) >= 0)
+    //    {
+    //        password = System.Web.HttpUtility.UrlEncode(password);
+    //    }
+
+    //    return password;
+    //}
+
+    private static readonly char[] UrlEncodeChars = { '+', '/', '=', '&' };
+
     public static string UrlEncodePassword(this string password)
     {
-        if (password != null && password.IndexOfAny("+/=&".ToCharArray()) >= 0)
+        if (string.IsNullOrEmpty(password))
+            return password;
+
+        // Use Span for efficient search
+        if (password.AsSpan().IndexOfAny(UrlEncodeChars) >= 0)
         {
             password = System.Web.HttpUtility.UrlEncode(password);
         }